pyproject.toml
[tool.poetry]
    name = "flask-api"
    version = "0.1.0"
    description = ""
    packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
    python = ">=3.12,<3.13"
    flask = ">=3.1.0,<4.0.0"
    flask-smorest = "^0.45.0"
    python-dotenv = "^1.1.0"
    spacy = "^3.8.5"
    torch = "2.2.1"
    groq = "^0.23.0"
    numpy = "^1.26"
pinecone = "^6.0.2"
sentence-transformers = "^4.1.0"

[build-system]
    requires = ["poetry-core>=2.0.0,<3.0.0"]
    build-backend = "poetry.core.masonry.api"